requirements.txt
flet>=0.21.0
psutil
structlog
pydantic
requests
PyJWT
py-machineid
